Moved AlertTab and StateHistory to app/features/alerting
